Use CC BY 4.0 as license instead of previous...
license. The previous license didn't allow people to sell the icons, but
CC BY 4.0 allows commercial use. If anyone wants to continue using the
previous license version of applauncher-icons, version 1.0 will still be
available. The only thing that version 2.0 of this project will change
is a better readme file and CC BY 4.0 licensing. Because of this, all
that people such as developers will need to do to use the version with
CC BY 4.0 is say that they're using applauncher-icons version 2.0.

Also, this commit improves the readme to explain what this project is
better.